вынес тесты к компонентам, добавил тест для <SmallMovieCard/> на mouseleave
