+++ src/components/small-movie-card/small-movie-card.e2e.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import {shallow} from "enzyme";
-import SmallMovieCard from "../small-movie-card.jsx";
+import SmallMovieCard from "./small-movie-card.jsx";
 
 const movie = {
   title: `Fantastic Beasts: The Crimes of Grindelwald`,
@@ -21,5 +21,10 @@ describe(`<SmallMovieCard />`, () => {
     wrapper.simulate(`mouseover`);
     expect(onMovieHover).toHaveBeenCalledTimes(1);
     expect(onMovieHover.mock.calls[0][0]).toMatchObject(movie);
+  });
+  it(`onMouseLeave`, () => {
+    wrapper.simulate(`mouseleave`);
+    expect(onMovieHover).toHaveBeenCalledTimes(2);
+    expect(onMovieHover.mock.calls[1][0]).toBe(null);
   });
 });
